requirements.txt
SpeechRecognition==3.8.1
requests==2.22.0
youtube_dl==2020.12.22
youtube_search_python==1.3.4
playsound==1.2.2
gTTS==2.2.1
python_vlc==3.0.11115
